% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTMap.R
\name{getTMap}
\alias{getTMap}
\title{Compute t-statistic map with or without Empirical Bayes.}
\usage{
getTMap(
  V,
  design,
  columnIndex = NULL,
  weights = NULL,
  brain.mask,
  eBayes = TRUE,
  output.file = "Tmap.nii.gz",
  returnObject = TRUE,
  writeToDisk = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{V}{Matrix of voxels intensities (rows are voxels, columns are
subjects).}

\item{design}{Design matrix for the covariates of interest, with rows
corresponding to subjects and columns to coefficients to be estimated.}

\item{columnIndex}{Numeric indicating which covariate of the design matrix
should be considered.}

\item{weights}{Non-negative observation weights. Can be a numeric matrix of
individual weivhts, of same size as the voxel matrix, or a numeric vector of
subject weights with length equal to 'ncol' of the voxel matrix, or a nmeric
vector of voxel weights with length equal to 'nrow' of the voxel matrix.}

\item{brain.mask}{Path of a NIfTI file mask to specify subset of voxels for
which the t-statistics should be computed.}

\item{eBayes}{Should the limma Empirical Bayes method be used to compute
moderated t-statistics?}

\item{output.file}{Filename of the NIfTI file to be saved to disk,
containing the computed t-statistics.}

\item{returnObject}{Should the t-statistics be returned as a matrix?}

\item{writeToDisk}{Should the t-statistics image be saved to the disk?}

\item{verbose}{Should messages be printed?}

\item{...}{additional arguments to pass to \code{\link{lmFit}}}
}
\value{
If \code{returnObject} is \code{TRUE}, a matrix of the t-statitics
is returned, and if \code{writeToDisk} is \code{TRUE}, the t-statistics
image is saved to disk as a NIfTI file.
}
\description{
Compute t-statistic map with or without Empirical Bayes.
}
\author{
Jean-Philippe Fortin
}
