% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks.R
\docType{data}
\name{landmarks}
\alias{landmarks}
\title{Estimated Landmarks for Various Imaging Modalities}
\format{
A list with 4 elements, which are \code{flair}, \code{pd}, \code{t1}, 
\code{t2}.  Each of these elements contains a matrix of 9 columns, each different
quantiles of the distribution of voxel intensites and 33 rows, corresponding to
healthy individuals
}
\usage{
landmarks
}
\description{
A list of histogram landmarks for histogram matching.
}
\keyword{datasets}
