\name{mask_intersect}
\alias{mask_intersect}
\title{
Intersection of a list of brain masks
}
\description{
 RAVEl stuff is blablablablabla. 
}
\usage{
mask_intersect(list, output.file = NULL, prob=1, reorient = FALSE, returnObject = TRUE,
	writeToDisk = TRUE, verbose = TRUE)
}
\arguments{
  \item{list}{
    Stuff
  }
}

\value{
No value returned. Instead corrected images are saved.
}
\author{
Jean-Philippe Fortin
}

