% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeZScore.R
\name{normalizeZScore}
\alias{normalizeZScore}
\title{Whole-brain z-score intensity normalization}
\usage{
normalizeZScore(
  input.files,
  output.files = NULL,
  brain.mask = NULL,
  writeToDisk = FALSE,
  returnMatrix = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{input.files}{Vector of filenames for the input images. Must be NIfTI
files.}

\item{output.files}{Optional vector of filenames for the output images. By
default, will be the \code{input.files} with suffix "ZScore".}

\item{brain.mask}{Filename for the brain binary mask specifying the template
space brain. Must be a NIfTI file.}

\item{writeToDisk}{Should the scans be saved to the disk? FALSE by default.}

\item{returnMatrix}{Should the matrix of intensities be returned? FALSE by default.}

\item{verbose}{Should messages be printed?}
}
\value{
if \code{returnMatrix} is \code{FALSE}, no value returned, but
Zscore-normalized images are saved. If \code{returnMatrix} is \code{TRUE},
Zscore-normalized images are saved and a matrix of normalized intensities is
returned.
}
\description{
Whole-brain z-score intensity normalization.
}
\author{
Jean-Philippe Fortin
}
