% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ravel-package}
\alias{ravel}
\alias{ravel-package}
\title{Ravel}
\description{
Ravel is an RStudio-native coding and analytics copilot for R users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/msaule/ravel}
  \item Report bugs at \url{https://github.com/msaule/ravel/issues}
}

}
\author{
\strong{Maintainer}: Markuss Saule \email{markusstomas@gmail.com} [copyright holder]

}
\keyword{internal}
