\name{RAVEL}
\alias{RAVEL}
\title{
RAVEL correction
}
\description{
 RAVEl stuff is blablablablabla. 
}
\usage{
RAVEL(input.files, output.files=NULL, brain.mask=NULL, control.mask=NULL, WhiteStripe=FALSE, WhiteStripe_Type="T1",  k=1, verbose=TRUE)
}
\arguments{
  \item{input.files}{
    Stuff
  }
  \item{output.files}{
    Stuff
  }
  \item{brain.mask}{
    Stuff
  }
  \item{control.mask}{
    Stuff
  }
  \item{WhiteStripe}{
    Should White Stripe intensity normalization be performed prior to RAVEL?
  }
  \item{WhiteStripe_Type}{
    Something
  }
  \item{k}{
    Number of unwanted factors to include in the RAVEL model
  }
  \item{verbose}{
    Something
  }
}

\value{
No value returned. Instead corrected images are saved.
}
\author{
Jean-Philippe Fortin
}

