% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approvals.R
\name{ravel_approve_action}
\alias{ravel_approve_action}
\title{Approve a staged action}
\usage{
ravel_approve_action(action)
}
\arguments{
\item{action}{A \code{ravel_action}.}
}
\value{
The updated action.
}
\description{
Approve a staged action
}
