% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ravel_get_setting}
\alias{ravel_get_setting}
\title{Get a Ravel setting}
\usage{
ravel_get_setting(key, default = NULL)
}
\arguments{
\item{key}{Setting name.}

\item{default}{Default value when the setting is missing.}
}
\value{
The setting value.
}
\description{
Get a Ravel setting
}
