% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_models.R
\name{ravel_interpret_model}
\alias{ravel_interpret_model}
\title{Interpret a model in plain English}
\usage{
ravel_interpret_model(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A character string.
}
\description{
Interpret a model in plain English
}
