% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_interface.R
\name{ravel_provider_capabilities}
\alias{ravel_provider_capabilities}
\title{Report provider capabilities}
\usage{
ravel_provider_capabilities(
  provider = c("openai", "copilot", "gemini", "anthropic")
)
}
\arguments{
\item{provider}{Provider name.}
}
\value{
A named list.
}
\description{
Report provider capabilities
}
