% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{ravel_read_history}
\alias{ravel_read_history}
\title{Read recent Ravel history entries}
\usage{
ravel_read_history(limit = 100L)
}
\arguments{
\item{limit}{Maximum number of history entries to return.}
}
\value{
A tibble.
}
\description{
Read recent Ravel history entries
}
