% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approvals.R
\name{ravel_reject_action}
\alias{ravel_reject_action}
\title{Reject a staged action}
\usage{
ravel_reject_action(action)
}
\arguments{
\item{action}{A \code{ravel_action}.}
}
\value{
The updated action.
}
\description{
Reject a staged action
}
