% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ravel_set_bearer_token}
\alias{ravel_set_bearer_token}
\title{Store a bearer token for a provider}
\usage{
ravel_set_bearer_token(provider = c("gemini"), token, persist = TRUE)
}
\arguments{
\item{provider}{Provider name.}

\item{token}{Bearer token value.}

\item{persist}{Whether to try to persist the token using \code{keyring}.}
}
\value{
The stored token, invisibly.
}
\description{
Store a bearer token for a provider
}
