% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ravel_set_setting}
\alias{ravel_set_setting}
\title{Set a Ravel setting}
\usage{
ravel_set_setting(key, value)
}
\arguments{
\item{key}{Setting name.}

\item{value}{Setting value.}
}
\value{
The written settings list, invisibly.
}
\description{
Set a Ravel setting
}
