% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_settings.R
\name{ravel_settings_addin}
\alias{ravel_settings_addin}
\title{Launch the Ravel settings addin}
\usage{
ravel_settings_addin()
}
\value{
Invisibly launches a Shiny gadget.
}
\description{
Launch the Ravel settings addin
}
