% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_settings.R
\name{ravel_setup_addin}
\alias{ravel_setup_addin}
\title{Launch the Ravel setup assistant}
\usage{
ravel_setup_addin()
}
\value{
Invisibly launches a Shiny gadget.
}
\description{
Launch the Ravel setup assistant
}
