\name{mask_intersect}
\alias{utils}
\title{
Intersection of a list of brain masks
}
\description{
 RAVEl stuff is blablablablabla. 
}
\usage{
mask_intersect(list)
}
\arguments{
  \item{list}{
    Stuff
  }
}

\value{
No value returned. Instead corrected images are saved.
}
\author{
Jean-Philippe Fortin
}

