% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ravel-package}
\alias{ravel}
\alias{ravel-package}
\title{Ravel}
\description{
Ravel is an RStudio-native coding and analytics copilot for R users.
}
\author{
\strong{Maintainer}: Ravel Contributors \email{opensource@ravel.dev}

}
\keyword{internal}
