% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{ravel_set_setting}
\alias{ravel_set_setting}
\title{Set a Ravel setting}
\usage{
ravel_set_setting(key, value)
}
\arguments{
\item{key}{Setting name.}

\item{value}{Setting value.}
}
\value{
The written settings list, invisibly.
}
\description{
Set a Ravel setting
}
\details{
Settings are stored in session memory by default. To mirror non-sensitive
settings to disk explicitly, configure \code{options(ravel.user_dirs = list( config = "<path>"))} before calling this function.
}
